Consolidate AuthenticatedUserRule with SetAndRestoreAuthenticatedUser
