+++ core/server/master/src/test/java/alluxio/master/file/PermissionCheckTest.java
@@ -17,7 +17,6 @@ import alluxio.ConfigurationRule;
 import alluxio.Constants;
 import alluxio.LoginUserRule;
 import alluxio.PropertyKey;
-import alluxio.SetAndRestoreAuthenticatedUser;
 import alluxio.SetAndRestoreConfiguration;
 import alluxio.exception.AccessControlException;
 import alluxio.exception.ExceptionMessage;
@@ -203,16 +202,14 @@ public final class PermissionCheckTest {
    */
   private void createDirAndFileForTest() throws Exception {
     // create "/testDir" for user1
-    try (SetAndRestoreAuthenticatedUser user = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_ADMIN.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_ADMIN.getUser()).toResource()) {
       mFileSystemMaster.createDirectory(new AlluxioURI("/testDir"),
           CreateDirectoryOptions.defaults().setOwner(TEST_USER_1.getUser())
               .setGroup(TEST_USER_1.getGroup()).setMode(TEST_DIR_MODE));
     }
 
     // create "/testDir/file" for user1
-    try (SetAndRestoreAuthenticatedUser user = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_1.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_1.getUser()).toResource()) {
       mFileSystemMaster.createFile(new AlluxioURI("/testDir/file"),
           CreateFileOptions.defaults().setBlockSizeBytes(Constants.KB)
               .setOwner(TEST_USER_1.getUser()).setGroup(TEST_USER_1.getGroup())
@@ -220,8 +217,7 @@ public final class PermissionCheckTest {
     }
 
     // create "/testFile" for user2
-    try (SetAndRestoreAuthenticatedUser user = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_ADMIN.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_ADMIN.getUser()).toResource()) {
       mFileSystemMaster.createFile(new AlluxioURI("/testFile"),
           CreateFileOptions.defaults().setBlockSizeBytes(Constants.KB)
               .setOwner(TEST_USER_2.getUser()).setGroup(TEST_USER_2.getGroup())
@@ -241,8 +237,7 @@ public final class PermissionCheckTest {
     permissions.add(new ImmutableTriple<>(TEST_USER_1.getUser(), TEST_USER_1.getGroup(),
         new Mode((short) 0754)));
     LockedInodePath lockedInodePath = getLockedInodePath(permissions);
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_1.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_1.getUser()).toResource()) {
       PermissionChecker checker = new PermissionChecker(mInodeTree);
       Mode.Bits actual = checker.getPermission(lockedInodePath);
       Assert.assertEquals(Mode.Bits.ALL, actual);
@@ -255,8 +250,7 @@ public final class PermissionCheckTest {
     permissions.add(new ImmutableTriple<>(TEST_USER_1.getUser(), TEST_USER_1.getGroup(),
         new Mode((short) 0754)));
     LockedInodePath lockedInodePath = getLockedInodePath(permissions);
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_3.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_3.getUser()).toResource()) {
       PermissionChecker checker = new PermissionChecker(mInodeTree);
       Mode.Bits actual = checker.getPermission(lockedInodePath);
       Assert.assertEquals(Mode.Bits.READ_EXECUTE, actual);
@@ -269,8 +263,7 @@ public final class PermissionCheckTest {
     permissions.add(new ImmutableTriple<>(TEST_USER_1.getUser(), TEST_USER_1.getGroup(),
         new Mode((short) 0754)));
     LockedInodePath lockedInodePath = getLockedInodePath(permissions);
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_2.getUser())) {
+    try (AutoCloseable r  = new AuthenticatedUserRule(TEST_USER_2.getUser()).toResource()) {
       PermissionChecker checker = new PermissionChecker(mInodeTree);
       Mode.Bits actual = checker.getPermission(lockedInodePath);
       Assert.assertEquals(Mode.Bits.READ, actual);
@@ -319,7 +312,7 @@ public final class PermissionCheckTest {
   }
 
   private void verifyCreateFile(TestUser user, String path, boolean recursive) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       CreateFileOptions options = CreateFileOptions.defaults().setRecursive(recursive)
           .setOwner(SecurityUtils.getOwnerFromThriftClient())
           .setGroup(SecurityUtils.getGroupFromThriftClient()).setPersisted(true);
@@ -374,7 +367,7 @@ public final class PermissionCheckTest {
 
   private void verifyCreateDirectory(TestUser user, String path, boolean recursive)
       throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       CreateDirectoryOptions options = CreateDirectoryOptions.defaults().setRecursive(recursive)
           .setOwner(SecurityUtils.getOwnerFromThriftClient())
           .setGroup(SecurityUtils.getGroupFromThriftClient());
@@ -449,7 +442,7 @@ public final class PermissionCheckTest {
   }
 
   private void verifyRename(TestUser user, String srcPath, String dstPath) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       String fileOwner =
           mFileSystemMaster.getFileInfo(mFileSystemMaster.getFileId(new AlluxioURI(srcPath)))
               .getOwner();
@@ -529,7 +522,7 @@ public final class PermissionCheckTest {
   }
 
   private void verifyDelete(TestUser user, String path, boolean recursive) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       mFileSystemMaster.delete(new AlluxioURI(path), DeleteOptions.defaults()
           .setRecursive(recursive));
       Assert.assertEquals(-1, mFileSystemMaster.getFileId(new AlluxioURI(path)));
@@ -582,8 +575,7 @@ public final class PermissionCheckTest {
     mThrown.expect(AccessControlException.class);
     mThrown.expectMessage(ExceptionMessage.PERMISSION_DENIED.getMessage(
         toExceptionMessage(TEST_USER_2.getUser(), Mode.Bits.READ, dir, "onlyReadByUser1")));
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_2.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_2.getUser()).toResource()) {
       verifyGetFileInfoOrList(TEST_USER_2, dir, false);
     }
   }
@@ -624,14 +616,14 @@ public final class PermissionCheckTest {
   }
 
   private void verifyRead(TestUser user, String path, boolean isFile) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       verifyGetFileId(user, path);
       verifyGetFileInfoOrList(user, path, isFile);
     }
   }
 
   private void verifyGetFileId(TestUser user, String path) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       long fileId = mFileSystemMaster.getFileId(new AlluxioURI(path));
       Assert.assertNotEquals(-1, fileId);
     }
@@ -639,7 +631,7 @@ public final class PermissionCheckTest {
 
   private void verifyGetFileInfoOrList(TestUser user, String path, boolean isFile)
       throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       if (isFile) {
         Assert.assertEquals(path, mFileSystemMaster.getFileInfo(new AlluxioURI(path)).getPath());
         Assert.assertEquals(1,
@@ -697,7 +689,7 @@ public final class PermissionCheckTest {
 
   private SetAttributeOptions verifySetState(TestUser user, String path,
       SetAttributeOptions options) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       mFileSystemMaster.setAttribute(new AlluxioURI(path), options);
 
       FileInfo fileInfo = mFileSystemMaster.getFileInfo(new AlluxioURI(path));
@@ -744,7 +736,7 @@ public final class PermissionCheckTest {
 
   private void verifyCompleteFile(TestUser user, String path, CompleteFileOptions options)
       throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       mFileSystemMaster.completeFile(new AlluxioURI(path), options);
     }
   }
@@ -796,7 +788,7 @@ public final class PermissionCheckTest {
   }
 
   private void verifyFree(TestUser user, String path, boolean recursive) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(user.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(user.getUser()).toResource()) {
       mFileSystemMaster.free(new AlluxioURI(path), FreeOptions.defaults().setRecursive(recursive));
     }
   }
@@ -900,14 +892,13 @@ public final class PermissionCheckTest {
 
   private void verifySetAcl(TestUser runUser, String path, String owner, String group,
       short mode, boolean recursive) throws Exception {
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(runUser.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(runUser.getUser()).toResource()) {
       SetAttributeOptions options =
           SetAttributeOptions.defaults().setOwner(owner).setGroup(group).setMode(mode)
               .setRecursive(recursive);
       mFileSystemMaster.setAttribute(new AlluxioURI(path), options);
     }
-    try (SetAndRestoreAuthenticatedUser u = new SetAndRestoreAuthenticatedUser(
-        TEST_USER_ADMIN.getUser())) {
+    try (AutoCloseable r = new AuthenticatedUserRule(TEST_USER_ADMIN.getUser()).toResource()) {
       FileInfo fileInfo =
           mFileSystemMaster.getFileInfo(mFileSystemMaster.getFileId(new AlluxioURI(path)));
       if (owner != null) {
